fix(kafka-outbound-connector): Fix serialization of request permission market documents
The serialization for Kafka used extra checks that were not updated to reflect the latest changes of the CIM serializer.
The serialization mechanism is replaced by a simpler adapter.
