--- outbound-connectors/outbound-kafka/src/main/java/energy/eddie/outbound/kafka/CustomSerializer.java
@@ -1,78 +1,0 @@
-// SPDX-FileCopyrightText: 2023-2026 The EDDIE Developers <eddie.developers@fh-hagenberg.at>
-// SPDX-License-Identifier: Apache-2.0
-
-package energy.eddie.outbound.kafka;
-
-import energy.eddie.cim.agnostic.ConnectionStatusMessage;
-import energy.eddie.cim.agnostic.OpaqueEnvelope;
-import energy.eddie.cim.agnostic.PermissionCommand;
-import energy.eddie.cim.agnostic.RawDataMessage;
-import energy.eddie.cim.serde.MessageSerde;
-import energy.eddie.cim.serde.SerializationException;
-import energy.eddie.cim.v0_82.ap.AccountingPointEnvelope;
-import energy.eddie.cim.v0_82.pmd.PermissionEnvelope;
-import energy.eddie.cim.v0_82.vhd.ValidatedHistoricalDataEnvelope;
-import energy.eddie.cim.v0_91_08.RTREnvelope;
-import energy.eddie.cim.v1_04.vhd.VHDEnvelope;
-import jakarta.annotation.Nullable;
-import org.apache.kafka.common.serialization.Serializer;
-import org.apache.kafka.common.serialization.StringSerializer;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-class CustomSerializer implements Serializer<Object> {
-    private static final Logger LOGGER = LoggerFactory.getLogger(CustomSerializer.class);
-    private final StringSerializer stringSerializer = new StringSerializer();
-    private final MessageSerde serde;
-
-    CustomSerializer(MessageSerde serde) {
-        this.serde = serde;
-    }
-
-    @Override
-    @Nullable
-    public byte[] serialize(String topic, Object data) {
-        return switch (data) {
-            // Agnostic
-            case ConnectionStatusMessage ignored -> serialize(data);
-            case RawDataMessage ignored -> serialize(data);
-            case OpaqueEnvelope ignored -> serialize(data);
-            case PermissionCommand ignored -> serialize(data);
-            // CIM v0.82
-            case ValidatedHistoricalDataEnvelope ignored -> serialize(data);
-            case PermissionEnvelope ignored -> serialize(data);
-            case AccountingPointEnvelope ignored -> serialize(data);
-            // CIM v0.91.08
-            case RTREnvelope ignored -> serialize(data);
-            // CIM v1.04
-            case VHDEnvelope ignored -> serialize(data);
-            case energy.eddie.cim.v1_04.rtd.RTDEnvelope ignored -> serialize(data);
-            // CIM v1.12
-            case energy.eddie.cim.v1_12.rtd.RTDEnvelope ignored -> serialize(data);
-            case energy.eddie.cim.v1_12.ack.AcknowledgementEnvelope ignored -> serialize(data);
-            case energy.eddie.cim.v1_12.recmmoe.RECMMOEEnvelope ignored -> serialize(data);
-            case null -> null;
-            default -> {
-                LOGGER.warn("Got invalid type to serialize {}", data.getClass());
-                yield null;
-            }
-        };
-    }
-
-    @Override
-    public void close() {
-        stringSerializer.close();
-    }
-
-    // Sonar wants us to return an empty array, but the kafka implementations return null, so will do the same
-    @SuppressWarnings("java:S1168")
-    @Nullable
-    private byte[] serialize(Object payload) {
-        try {
-            return serde.serialize(payload);
-        } catch (SerializationException e) {
-            LOGGER.warn("Could not serialize message of type {}", payload.getClass(), e);
-            return null;
-        }
-    }
-}
