+++ outbound-connectors/outbound-kafka/src/main/java/energy/eddie/outbound/kafka/KafkaOutboundConnector.java
@@ -192,7 +192,7 @@ public class KafkaOutboundConnector {
             MessageSerde serde
     ) {
         var config = kafkaProperties(kafkaProperties);
-        return new DefaultKafkaProducerFactory<>(config, new StringSerializer(), new CustomSerializer(serde));
+        return new DefaultKafkaProducerFactory<>(config, new StringSerializer(), new KafkaSerializerAdapter(serde));
     }
 
     @Bean
